+++ lib/slack/block_kit/composition/option_group.rb
@@ -17,8 +17,8 @@ module Slack
           yield(self) if block_given?
         end
 
-        def option(text:, value:, emoji: nil)
-          @options << Option.new(text: text, value: value, emoji: emoji)
+        def option(text:, value:, emoji: nil, initial: false)
+          @options << Option.new(text: text, value: value, emoji: emoji, initial: initial)
 
           self
         end